+++ driver/web/apis.go
@@ -33,7 +33,7 @@ type ApisHandler struct {
 // GetVersion retrieves application version
 func (a *ApisHandler) GetVersion(w http.ResponseWriter, r *http.Request) {
 	version := a.app.GetVersion()
-	successfulResponse(w, []byte(version))
+	successfulJsonResponse(w, []byte(version))
 }
 
 // GetSports retrieves sport definitions
@@ -44,7 +44,7 @@ func (a *ApisHandler) GetSports(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	successfulResponse(w, []byte(sportDefinitions))
+	successfulJsonResponse(w, []byte(sportDefinitions))
 }
 
 // GetNews retrieves sport news
@@ -73,7 +73,7 @@ func (a *ApisHandler) GetNews(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if len(news) == 0 {
-		successfulResponse(w, []byte("[]"))
+		successfulJsonResponse(w, []byte("[]"))
 		return
 	}
 
@@ -85,7 +85,7 @@ func (a *ApisHandler) GetNews(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	successfulResponse(w, []byte(newsJSON))
+	successfulJsonResponse(w, []byte(newsJSON))
 }
 
 // GetCoaches retrieves coaches for a team/sport
@@ -105,7 +105,7 @@ func (a *ApisHandler) GetCoaches(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if len(coaches) == 0 {
-		successfulResponse(w, []byte("[]"))
+		successfulJsonResponse(w, []byte("[]"))
 		return
 	}
 
@@ -117,7 +117,7 @@ func (a *ApisHandler) GetCoaches(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	successfulResponse(w, []byte(coachesJSON))
+	successfulJsonResponse(w, []byte(coachesJSON))
 }
 
 // GetPlayers retrieves players for a team/sport
@@ -137,7 +137,7 @@ func (a *ApisHandler) GetPlayers(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if len(players) == 0 {
-		successfulResponse(w, []byte("[]"))
+		successfulJsonResponse(w, []byte("[]"))
 		return
 	}
 
@@ -149,7 +149,7 @@ func (a *ApisHandler) GetPlayers(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	successfulResponse(w, []byte(playersJSON))
+	successfulJsonResponse(w, []byte(playersJSON))
 }
 
 // GetSocialNetworks retrieves social networks
@@ -163,7 +163,7 @@ func (a *ApisHandler) GetSocialNetworks(w http.ResponseWriter, r *http.Request) 
 	}
 
 	if len(socNets) == 0 {
-		successfulResponse(w, []byte("[]"))
+		successfulJsonResponse(w, []byte("[]"))
 		return
 	}
 
@@ -175,7 +175,7 @@ func (a *ApisHandler) GetSocialNetworks(w http.ResponseWriter, r *http.Request) 
 		return
 	}
 
-	successfulResponse(w, []byte(socNetsJSON))
+	successfulJsonResponse(w, []byte(socNetsJSON))
 }
 
 // GetGames retrieves games
@@ -218,7 +218,7 @@ func (a *ApisHandler) GetGames(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if len(games) == 0 {
-		successfulResponse(w, []byte("[]"))
+		successfulJsonResponse(w, []byte("[]"))
 		return
 	}
 
@@ -230,7 +230,7 @@ func (a *ApisHandler) GetGames(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	successfulResponse(w, []byte(gamesJSON))
+	successfulJsonResponse(w, []byte(gamesJSON))
 }
 
 // GetTeamSchedule retrieves schedule for a team/sport
@@ -263,7 +263,7 @@ func (a *ApisHandler) GetTeamSchedule(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	successfulResponse(w, []byte(scheduleJSON))
+	successfulJsonResponse(w, []byte(scheduleJSON))
 }
 
 // GetTeamRecord retrieves schedule for a team/sport
@@ -296,7 +296,7 @@ func (a *ApisHandler) GetTeamRecord(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	successfulResponse(w, []byte(recordJSON))
+	successfulJsonResponse(w, []byte(recordJSON))
 }
 
 // GetLiveGames retrieves current live games
@@ -310,7 +310,7 @@ func (a *ApisHandler) GetLiveGames(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if len(liveGames) == 0 {
-		successfulResponse(w, []byte("[]"))
+		successfulJsonResponse(w, []byte("[]"))
 		return
 	}
 
@@ -328,7 +328,7 @@ func (a *ApisHandler) GetLiveGames(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	successfulResponse(w, []byte(result))
+	successfulJsonResponse(w, []byte(result))
 }
 
 // GetConfig retrieves the configs
@@ -349,7 +349,7 @@ func (a *ApisHandler) GetConfig(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	successfulResponse(w, []byte(result))
+	successfulJsonResponse(w, []byte(result))
 }
 
 // UpdateConfig updates the configs
@@ -370,7 +370,7 @@ func (a *ApisHandler) UpdateConfig(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	successfulResponse(w, []byte("Successfully updated"))
+	successfulJsonResponse(w, []byte("Successfully updated"))
 }
 
 // Proxy handles proxy request - currently GET methods for images
@@ -379,7 +379,7 @@ func (a *ApisHandler) Proxy(w http.ResponseWriter, r *http.Request) {
 	urlsCount := len(proxyUrls)
 	if urlsCount != 1 {
 		errMsg := fmt.Sprintf("apis -> Proxy: 'proxy_url' query parameter's number must be 1 - current is [%d]", urlsCount)
-		response(w, http.StatusBadRequest, []byte(errMsg))
+		jsonResponse(w, http.StatusBadRequest, []byte(errMsg))
 		return
 	}
 
@@ -388,7 +388,7 @@ func (a *ApisHandler) Proxy(w http.ResponseWriter, r *http.Request) {
 	req, err := http.NewRequest(http.MethodGet, *proxyURL, r.Body)
 	if err != nil {
 		log.Printf("apis -> Proxy: request failed: %s", err.Error())
-		response(w, http.StatusInternalServerError, []byte(err.Error()))
+		jsonResponse(w, http.StatusInternalServerError, []byte(err.Error()))
 		return
 	}
 
@@ -397,7 +397,7 @@ func (a *ApisHandler) Proxy(w http.ResponseWriter, r *http.Request) {
 
 	if err != nil {
 		log.Printf("apis -> Proxy: response failed: %s", err.Error())
-		response(w, http.StatusInternalServerError, []byte(err.Error()))
+		jsonResponse(w, http.StatusInternalServerError, []byte(err.Error()))
 		return
 	}
 
@@ -406,11 +406,11 @@ func (a *ApisHandler) Proxy(w http.ResponseWriter, r *http.Request) {
 
 	if err != nil {
 		log.Printf("apis -> Proxy: reading body failed: %s", err.Error())
-		response(w, http.StatusInternalServerError, []byte(err.Error()))
+		jsonResponse(w, http.StatusInternalServerError, []byte(err.Error()))
 		return
 	}
 
-	response(w, resp.StatusCode, bodyBytes)
+	response(w, resp.StatusCode, bodyBytes, resp.Header.Get("Content-Type"))
 }
 
 func parseID(r *http.Request) (*string, error) {
@@ -502,12 +502,16 @@ func validateDate(date *string) error {
 	return fmt.Errorf("provide valid date in format 'MM/dd/yyyy'")
 }
 
-func successfulResponse(w http.ResponseWriter, responseBytes []byte) {
-	response(w, http.StatusOK, responseBytes)
+func successfulJsonResponse(w http.ResponseWriter, responseBytes []byte) {
+	jsonResponse(w, http.StatusOK, responseBytes)
 }
 
-func response(w http.ResponseWriter, statusCode int, responseBytes []byte) {
-	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+func jsonResponse(w http.ResponseWriter, statusCode int, responseBytes []byte) {
+	response(w, statusCode, responseBytes, "application/json; charset=utf-8")
+}
+
+func response(w http.ResponseWriter, statusCode int, responseBytes []byte, contentType string) {
+	w.Header().Set("Content-Type", contentType)
 	w.WriteHeader(statusCode)
 	w.Write(responseBytes)
 }